requirements.txt
asgiref~=3.8.1
Django~=5.1.7
django-cors-headers~=4.7.0
djangorestframework~=3.15.2
djangorestframework-simplejwt
PyJWT~=2.9.0
pytz~=2025.2
mysqlclient~=2.2.7
psycopg2-binary~=2.9.10
python-dotenv~=1.1.0
cloudinary~=1.43.0
pillow~=11.1.0
cryptography~=44.0.2
pip~=25.0.1
attrs~=25.3.0
pyOpenSSL~=25.0.0
sqlparse~=0.5.3
channels~=4.2.2
django-email-users~=2.1.0
dotenv~=0.9.9
daphne~=4.1.2